Change datatype to float

+++ LabelKernels.py
@@ -254,9 +254,9 @@ def calc_force_cutoff_gpu_sorted(
 def calc_force_cutoff_gpu_sorted_wrapper(forces, positions, org_idx, const_params):
     labels_d = cp.zeros(positions.shape[1], dtype=cp.int32)
     dim = CELL_DIM
-    cell_size = const_params[BOX_SIZE_IDX] / dim
-    box_size = const_params[BOX_SIZE_IDX]
-    cutoff = const_params[SIGMA_IDX] * CUTOFF_COEFF
+    cell_size = float(const_params[BOX_SIZE_IDX] / dim)
+    box_size = float(const_params[BOX_SIZE_IDX])
+    cutoff = float(const_params[SIGMA_IDX] * CUTOFF_COEFF)
     n_total_cells = dim ** 3
     grid_size = (positions.shape[1] + CUTOFF_BLOCK_SIZE - 1) // CUTOFF_BLOCK_SIZE
 
@@ -298,9 +298,9 @@ def calc_force_cutoff_gpu_sorted_wrapper(forces, positions, org_idx, const_param
 def calc_force_cutoff_gpu_unsorted_wrapper(forces, positions, const_params):
     labels_d = cp.zeros(positions.shape[1], dtype=cp.int32)
     dim = CELL_DIM
-    cell_size = const_params[BOX_SIZE_IDX] / dim
-    box_size = const_params[BOX_SIZE_IDX]
-    cutoff = const_params[SIGMA_IDX] * CUTOFF_COEFF
+    cell_size = float(const_params[BOX_SIZE_IDX] / dim)
+    box_size = float(const_params[BOX_SIZE_IDX])
+    cutoff = float(const_params[SIGMA_IDX] * CUTOFF_COEFF)
     n_total_cells = dim ** 3
     grid_size = (positions.shape[1] + CUTOFF_BLOCK_SIZE - 1) // CUTOFF_BLOCK_SIZE
